chore: cleanup unused imports etc. refactor: move paginationBar state out of its parent into the PaginationBar component.
