+++ src/components/Pagination/styles.js
@@ -1,7 +1,7 @@
 import styled from "styled-components";
 import { buttonBaseStyle } from "components/shared/Button";
 
-export const StyledPaginationButton = styled.a`
+export const StyledPaginationButton = styled.button`
     ${buttonBaseStyle}
     color: ${({theme}) => theme.colors.primary};
     background-color: ${({theme}) => theme.colors.lightest};
@@ -12,17 +12,17 @@ export const StyledPaginationButton = styled.a`
         text-decoration: none;
         color: inherit;
     }
+    &[disabled] {
+        filter: opacity(50%) grayscale(100%);
+        cursor: not-allowed;
+    }
 
     ${({active, theme}) => active ? 
     `
         background-color: ${theme.colors.primary};
         color: ${theme.colors.lightest};
-
-        &:hover {
-
-        }
     `: `
-        &:hover {
+        &:hover[enabled] {
             background-color: ${theme.colors.light};
             filter: none;
         }
