+++ src/components/SearchResult/PaginationBar.js
@@ -20,6 +20,9 @@ import { useMediaQuery } from 'react-responsive';
 import styled from 'styled-components';
 import { StyledLabeledTextInput } from 'components/shared/input/TextInput';
 import { Item, Next, Pagination, Prev } from 'components/Pagination';
+import { useEffect, useState } from 'react';
+import { useDispatch } from 'react-redux';
+import { updateSearchPageParams } from 'features/searchPageParamsReducer';
 
 const Container = styled.div`
     display: flex;
@@ -42,7 +45,51 @@ const PagePosition = styled.div`
     white-space: nowrap;
 `
 
-const PaginationBar = ({pageCount, currentPageIndex, goToPage, searchPageParams, setPageSize}) => {
+const PaginationBar = ({searchResults, searchPageParams}) => {
+
+    const [pageCount, setPageCount] = useState(0);
+    const [currentPageIndex, setCurrentPageIndex] = useState(0);
+    const dispatch = useDispatch();
+
+    useEffect(() => {
+        if(!searchResults){
+            setPageCount(0);
+            return;
+        }
+        if(searchPageParams.from === 0) {
+            setCurrentPageIndex(0);
+        }
+        let newPageCount = Math.ceil(searchResults.hitCount / searchPageParams.size);
+        setPageCount(newPageCount);
+    }, [searchResults, searchPageParams.size, searchPageParams.from])
+
+    useEffect(() => {
+        const from = currentPageIndex * searchPageParams.size;
+        dispatch(updateSearchPageParams({...searchPageParams, from}));
+        // TODO: Investigate / fix the circular looping here and remove es-lint disable
+        // eslint-disable-next-line 
+    }, [currentPageIndex, dispatch, searchPageParams.size]);
+
+    const goToPage = (pageNumber) => {
+        if(pageNumber >= 0) {
+            setCurrentPageIndex(pageNumber);
+        };
+    }
+
+    /**
+     * Handles input in hits per page field and rejets any
+     * value < 1 and > 999, i.e. leading zeros are also rejected.
+     */
+    const setPageSize = (e) => {
+        const re = /^[0-9\b]{1,3}$/;
+        if (e.target.value === '' || re.test(e.target.value)) {
+            let pageCount = parseInt(e.target.value);
+            if(pageCount === 0){
+                return;
+            }
+            dispatch(updateSearchPageParams({...searchPageParams, size: e.target.value}))
+        } 
+    }
 
     const isMobile = useMediaQuery({ query: '(max-width: 539px)' })
 
@@ -92,7 +139,7 @@ const PaginationBar = ({pageCount, currentPageIndex, goToPage, searchPageParams,
                         />
                         {renderPaginationItems()}
                         <Next onClick={() => goToPage(currentPageIndex + 1)}
-                            disabled={currentPageIndex === pageCount - 1}
+                            disabled={currentPageIndex + 1 === pageCount}
                         />
                     </Pagination>
                     <PagePosition>{`${currentPageIndex + 1} / ${pageCount}`}</PagePosition>
