+++ src/components/SearchResult/SearchResultList.js
@@ -15,13 +15,10 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
  */
-import {useState, useEffect} from 'react';
 import SearchResultItem from './SearchResultItem';
 import LoadingOverlay from 'components/shared/LoadingOverlay';
 import SearchBox from './SearchBox';
 import PaginationBar from './PaginationBar';
-import { useDispatch } from 'react-redux';
-import { updateSearchPageParams } from 'features/searchPageParamsReducer';
 import styled from 'styled-components';
 import { mobile } from 'config/media';
 import { sortLogsDateCreated } from 'utils';
@@ -82,48 +79,6 @@ const SearchResultList = ({
     className
 }) => {
 
-    const [pageCount, setPageCount] = useState(0);
-    const [currentPageIndex, setCurrentPageIndex] = useState(0);
-    const dispatch = useDispatch();
-
-    useEffect(() => {
-        if(!searchResults){
-            setPageCount(0);
-            return;
-        }
-        if(searchPageParams.from === 0) {
-            setCurrentPageIndex(0);
-        }
-        let newPageCount = Math.ceil(searchResults.hitCount / searchPageParams.size);
-        setPageCount(newPageCount);
-    }, [searchResults, searchPageParams.size, searchPageParams.from])
-
-    useEffect(() => {
-        const from = currentPageIndex * searchPageParams.size;
-        dispatch(updateSearchPageParams({...searchPageParams, from}));
-    }, [currentPageIndex, dispatch, updateSearchPageParams, searchPageParams.size]);
-
-    const goToPage = (pageNumber) => {
-        if(pageNumber >= 0) {
-            setCurrentPageIndex(pageNumber);
-        };
-    }
-
-    /**
-     * Handles input in hits per page field and rejets any
-     * value < 1 and > 999, i.e. leading zeros are also rejected.
-     */
-    const setPageSize = (e) => {
-        const re = /^[0-9\b]{1,3}$/;
-        if (e.target.value === '' || re.test(e.target.value)) {
-            let pageCount = parseInt(e.target.value);
-            if(pageCount === 0){
-                return;
-            }
-            dispatch(updateSearchPageParams({...searchPageParams, size: e.target.value}))
-        } 
-    }
-
     // Guarantee that sort is applied for the current page of results
     const sortedLogs = [...searchResults.logs];
     if(sortedLogs.length > 0) {
@@ -148,7 +103,7 @@ const SearchResultList = ({
                 <SearchResultsContainer id='search-results-list--container' aria-label='Search Results' >
                     {renderedSearchResults}
                 </SearchResultsContainer>
-                <StyledPaginationBar {...{pageCount, currentPageIndex, goToPage, searchPageParams, setPageSize}} />
+                <StyledPaginationBar {...{searchResults, searchPageParams}} />
             </StyledLoadingOverlay>
         </Container>
     )
